requirements.txt
###### Requirements with Version Specifiers ######
firebase_functions~=0.1.0
functions-framework==3.*
###### Requirements without Version Specifiers ######
alpaca-py
setuptools
polygon-api-client
ratelimit